plugins/requirements.txt
xmltodict
requests
pyjwt
protobuf==3.20
pyperclip

